requirements.txt
numpy
scipy
freud-analysis
fsph
